Make `RecordType` `#[non_exhaustive]`
This change ensures that adding new record types in the future will not
result in breaking changes for downstream users who have exhaustively
matched on the `RecordType` variants. This makes it easier for both the
library maintainers and users to evolve the codebase and adapt to new
requirements without introducing breaking changes.